requirements.txt
django
django-filter
python-dotenv
djangorestframework